Restructure opf mod files
