+++ src/gurobi_optimods/opf/__init__.py
@@ -1,0 +1,6 @@
+from gurobi_optimods.opf.api import (
+    solve_opf_model,
+    read_coords_from_csv_file,
+    read_case_from_mat_file,
+    turn_solution_into_mat_file,
+)
